+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - (Improvement) Block traffic on the services if there is more than 1 active leader in ActiveFailover mode
 - (Improvement) Improve master endpoint validation.
 - (Feature) Agency Improvements
+- (Bugfix) Fix agency timeout
+- (Improvement) Extract Agency Timeout
 
 ## [1.2.30](https://github.com/arangodb/kube-arangodb/tree/1.2.30) (2023-06-16)
 - (Feature) AgencyCache Interface